feat(server): add better-auth configuration and extract refreshAuthorizedEstablishments
# Conflicts:
#	server/src/controllers/auth-controller.ts
